Version 0.89
Changed Error Handler to Sentry / Raven
1)	Order Details – Vendor always being set to TBD on save.
a.	This error was caused when we put in the new code that reset the Tax to unchecked if the user selected Customer Provided.
b.	I had to change the way this was done and it will be working in version 0.89
2)	Copying a Line – Carries over phantom numbers.
a.	This was found after 0.88 was released.
b.	I put a change in that now when a line is copied, the quantity and totals are cleared.  Price each is still copied over.
3)	Copying a Line – Odd Extra Line
a.	Still trying to recreate this. Any ideas would be helpful.
4)	Save of Order swaps Art Placement Intermittently.
a.	Fixed it to sort by internal order_art_placement_id in ascending order.
b.	This will put the first added art placement at the top of the list and the others below it in the correct sequential order.
5)	Garment Screen – Vendor Drop down allowing for Typing without filtering.
a.	I verified that the value typed in does not save to the database.  The old value is used if the typed in value does not exist.
b.	This functionality is built into the drop down control and cannot be changed.
c.	If someone selects the entire text (CTRL + A) then types, it will filter as expected.  This is a training point unless we want to rewrite the entire garments screen.
d.	I’m going to keep doing research to see if there is anything else to do.
e.	The other option that I have enabled right now is that you can’t type in the drop down to filter the list.
6)	Reorders saved as a Quote does not update the Order Status History
a.	This is caused by the fix to not perform a order status history update on the initial reorder.
b.	I have implemented it so the order status is not set when a reorder takes place.
c.	Because it is required.  The screen will not allow the user to save until they select the status that they want.
7)	Date Changing on Save of Reorder
a.	I reordered 5 times and could not get this to happen.